+++ src/frontend/apps/impress/src/pages/docs/[id]/index.tsx
@@ -19,6 +19,7 @@ import {
   useTrans,
 } from '@/docs/doc-management/';
 import { KEY_AUTH, setAuthUrl, useAuth } from '@/features/auth';
+import { FloatingBar } from '@/features/docs/doc-header/components/FloatingBar';
 import { getDocChildren, subPageToTree } from '@/features/docs/doc-tree/';
 import { DocEditorSkeleton, useSkeletonStore } from '@/features/skeletons';
 import { MainLayout } from '@/layouts';
@@ -60,6 +61,7 @@ export function DocLayout() {
         }}
       >
         <MainLayout enableResizablePanel={true}>
+          <FloatingBar />
           <DocPage id={id} />
         </MainLayout>
       </TreeProvider>